Prepare for the release of Xcode 4, which has working LTO with static libraries.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@7280 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/HW/DSPHLE/UCodes/UCode_Zelda_Voice.cpp
@@ -235,7 +235,7 @@ void PrintObject(const T &Obj)
 	CompileTimeAssert<sizeof(ZeldaVoicePB) == 0x180> ensure_zpb_size_correct;
 	(void)ensure_zpb_size_correct;
 
-	for(int i = 0; i < sizeof(T); i++) {
+	for (size_t i = 0; i < sizeof(T); i++) {
 		if((i > 0) && ((i & 1) == 0))
 			ss << " ";
